requirements.txt
Pillow==8.2.0
PyQt5==5.15.2
pyinstaller==4.7